+++ src/Charts/GanttChart.php
@@ -3,13 +3,14 @@
 namespace Khill\Lavacharts\Charts;
 
 /**
- * TimelineChart Class
+ * GanttChart Class
  *
- *
- * A timeline is a chart that depicts how a set of resources are used over time.
- * If you're managing a software project and want to illustrate who is doing what
- * and when, or if you're organizing a conference and need to schedule meeting
- * rooms, a timeline is often a reasonable visualization choice.
+ * A Gantt chart is a type of chart that illustrates the breakdown of a
+ * project into its component tasks. Google Gantt charts illustrate the
+ * start, end, and duration of tasks within a project, as well as any
+ * dependencies a task may have. Google Gantt charts are rendered in
+ * the browser using SVG. Like all Google charts, Gantt charts display
+ * tooltips when the user hovers over the data.
  *
  *
  * @package    Khill\Lavacharts
@@ -21,14 +22,14 @@ namespace Khill\Lavacharts\Charts;
  * @link       http://lavacharts.com                   Official Docs Site
  * @license    http://opensource.org/licenses/MIT MIT
  */
-class TimelineChart extends Chart
+class GanttChart extends Chart
 {
     /**
      * Javascript chart type.
      *
      * @var string
      */
-    const TYPE = 'TimelineChart';
+    const TYPE = 'GanttChart';
 
     /**
      * Javascript chart version.
@@ -42,12 +43,12 @@ class TimelineChart extends Chart
      *
      * @var string
      */
-    const VIZ_PACKAGE = 'timeline';
+    const VIZ_PACKAGE = 'gantt';
 
     /**
      * Google's visualization class name.
      *
      * @var string
      */
-    const VIZ_CLASS = 'google.visualization.Timeline';
+    const VIZ_CLASS = 'google.visualization.Gantt';
 }