+++ src/Configs/EventManager.php
@@ -35,7 +35,7 @@ class EventManager implements \Countable, \IteratorAggregate
      *
      * @var array
      */
-    private static $defaultEvents = [
+    private static $EVENT_TYPES = [
         'animationfinish',
         'error',
         'mouseout',
@@ -126,8 +126,8 @@ class EventManager implements \Countable, \IteratorAggregate
      */
     private function validEvent($event)
     {
-        if (in_array($event, static::$defaultEvents) === false) {
-            throw new InvalidEvent($event, static::$defaultEvents);
+        if (in_array($event, static::$EVENT_TYPES) === false) {
+            throw new InvalidEvent($event, static::$EVENT_TYPES);
         }
     }
 }